refact: 更新 MikanFetcher 模块，优化动画列表和详情页抓取逻辑，增强数据解析能力，统一数据类型为 i64，提升代码一致性和可读性。

+++ src-tauri/src/services/crawler_service.rs
@@ -76,7 +76,7 @@ impl CrawlerService {
         };
         
         let fetcher = MikanFetcher::new(base_url, proxy);
-        let limit = params.limit.map(|v| v as usize);
+        let limit = params.limit.map(|v| v as i64);
         let detail_urls = match fetcher.fetch_and_parse_list(&list_url, limit).await {
             Ok(urls) => {
                 self.total_items = urls.len() as i64;
